BL-747 Item limits (#35)
Currently item limits in requests are making it impossible for patrons 
to request items that come from large sets of records since it is 
limited to the first 100 records.  This change is similar to what was 
implemented for loans and request sets so that we can access all of the 
records in the request dropdowns.