Always retrieve the max limit from collection api.
This commit updates how `Alma::Electronic.get_ids` retrieves all the ids
in the set collections.

Before this change if it were determined that there are a total of 101
collections, then the method would make two requests the first to get
100 items and the second to get only 1 item.

With this change we will still make two requests, but in each case the
limit will be set to the max allowed, 100.

The reason for this change is to compensate for what looks to be a bug in
the alma/electronic API. When retrieving all the collections the
last record of the first 100 collections is always the same as the first record
of the next 100. This does not happen for any of the other request pair (a
total of 23)

There's a bug in alma/electronic api where one collection gets duplicated:

The last collection of the first 100 is always the same the first
collection of next 100.