Make sure fee is always the same type.
Usually if the user has no fees/fines it is returned as 0.0, but in case of
nil we have been setting it to a string ('0').

By defaulting to the 0.0 value  instead of a '0' string we can reliably
count on using `user.total_fines > 0`  or `user.total_fines > 0.0` which
would fail if `user.total_fines == '0'`.